Introductory sentence introducing short cuts
Currently a lot of short cuts are used in the Quickstart document.
Even though there might be a glossary somewhere explaining them
all the document is hard to read for people just willing to give
the quickstart a go.

This commit adds an initial sentence properly introducing the
short cuts so that the quickstart document should be easier to
comprehend.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>