create yaml files
Currently the generated yamls are directly piped into kubectl.
The text however states that a file is created.

This commit creates files and uses kubectl to apply the files
instead of applying the yamls directly into the cluster using a pipe.
As one can have a different setup and make mistakes, it is best to have
the files in place to be able to easily correct and reproduce results.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>