highlight configuration options
Currently the yaml files are very static and
bound to a single cluster stack.
It is not easy to see which line should be
adapted and how if another stack were to be used.

This commit introduces shell variables to
highlight the configuration options and give a
first hint on what can be configured and how.
Additionally it allows for easy namespace (read: tenant)
changes.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>