:book: Improve quickstart guide (#99)
* use heading level 1 (#) for first heading

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>

* remove prerequisite envsubst

Currently the installaiton of envsubst is documented redundantly.
The installation of envsubst is explained later in the document,
yet it is listed as a prerequisite.

This commit removed the envsubst installation from the prerequisites list.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>

* minor format fixes

- remove end of line spaces
- write pods instead of po

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>

* create yaml files

Currently the generated yamls are directly piped into kubectl.
The text however states that a file is created.

This commit creates files and uses kubectl to apply the files
instead of applying the yamls directly into the cluster using a pipe.
As one can have a different setup and make mistakes, it is best to have
the files in place to be able to easily correct and reproduce results.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>

* use current version of CSP helper

Currently the CSP helper version is hardcoded to an old version.

This commit adds the use of curl to determine the latest version
of the csp helper (as already used with the operator deployments).

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>

* highlight configuration options

Currently the yaml files are very static and
bound to a single cluster stack.
It is not easy to see which line should be
adapted and how if another stack were to be used.

This commit introduces shell variables to
highlight the configuration options and give a
first hint on what can be configured and how.
Additionally it allows for easy namespace (read: tenant)
changes.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>

---------

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>
Co-authored-by: paul <hildebrandt@b1-systems.de>