fix(ci): Fix .yamllint file
Current yamllint validation fails (hmm, how was the check introduced if
it is failing). Fix this by explicitly excluding helm templated paths
from the check since it is NOT possible to check them with yamllint
(there is helm lint verification instead). In addition to that skip
`document-start` check until all files get the required header.

Signed-off-by: Artem Goncharov <artem.goncharov@gmail.com>