fix(cluster-class): Remove workload_loadbalancer variable
The workload_loadbalancer should configure the lb-provider used in the
Workload Cluster by overwriting the values for OCCM using the addons
controller of the Cluster Stack Operator.
Since we create the cloud-config for the OCCM using the
openstack-csp-helper we need to disable the creation of the
cluster-config secret using the OCCM Helm chart. With that no change of
the cloud-config using the OCCM values will be applies, so the
workload_loadbalancer variable has no effect.

Signed-off-by: Jan Schoone <jan.schoone@uhurutec.com>